Move function to stumpy.core
